Use explicitly on format warnings for Time test
It looks like the compile check doesn't work properly on some
architectures, which appears to eat or ignore the -Werror flag.

Instead, be more specific in that we state that printf formatting errors
are to be triggered on, and treated as an error explicitly, which in the
end is exactly what we are after.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>